# test: speed up trigger update tests with mock time provider