## Summary
- Implemented TimeProvider interface to abstract time operations
- Added mock time provider for testing, reducing test execution time from ~10.2s to ~0.2s (50x speedup)
- Maintains test reliability while significantly improving developer experience

## Background
The trigger update tests were taking over 5 seconds each because they needed to wait for the actual UpdateDelay configured for HomeAirConditioner operation mode changes. This made the test suite slow and impacted development velocity.

## Changes
1. **TimeProvider Interface**: Created an abstraction for time-related operations
2. **RealTimeProvider**: Production implementation using standard time.After()
3. **MockTimeProvider**: Test implementation with manual time control via Advance() method
4. **WebSocketServer Integration**: Injected TimeProvider dependency
5. **Test Updates**: Modified trigger update tests to use mock time

## Test Results
- TestSetPropertiesWithTriggerUpdate: 5.10s → 0.06s (~85x faster)
- TestSetPropertiesWithTriggerUpdateTiming: 5.10s → 0.12s (~42x faster)
- Total test time reduction: 10.2s → 0.18s

## Test Plan
- [x] All existing tests pass
- [x] Mock time tests are stable (ran 5 times consecutively)
- [x] Production behavior unchanged (using RealTimeProvider by default)
- [x] No changes to business logic, only test infrastructure

🤖 Generated with [Claude Code](https://claude.ai/code)